:recycle: Move developer.dart to utils/developer.dart for better organization
